+ Correct rating will be reduced from loosing arena team. + Mana regen. related patch ported from rev.2471 .

+++ trunk_2319/src/game/Arenas.cpp
@@ -337,13 +337,11 @@ void Arena::Finish()
 			//                   (PB - PA)/400
 			//              1 + 10
 
-			long double power = ( averageRating[j] - averageRating[i] ) / 400.0;
-
-			long double divisor = pow(((long double)(10.0)), power);
+			double power = (int)(averageRating[j] - averageRating[i]) / 400.0f;
+			double divisor = pow(((double)(10.0)), power);
 			divisor += 1.0;
 
-			long double winChance = 1.0 / divisor;
-
+			double winChance = 1.0 / divisor;
 			for(set<Player*>::iterator itr = m_players[i].begin(); itr != m_players[i].end(); ++itr)
 			{
 				Player * plr = *itr;
@@ -364,8 +362,8 @@ void Arena::Finish()
 						// K is the maximum possible change
 						// Through investigation, K was estimated to be 30
 						// When used in chess, Elo uses K = 32 ... maybe this would be better? :/
-						long double multiplier = (outcome ? 1.0 : 0.0) - winChance;
-						long double deltaRating = 30.0 * multiplier;
+						double multiplier = (outcome ? 1.0 : 0.0) - winChance;
+						double deltaRating = 30.0 * multiplier;
 						if ( deltaRating < 0 && (-1.0 * deltaRating) > t->m_stat_rating )
 							t->m_stat_rating = 0;
 						else
